cap size_t length diff safely
Agent-Logs-Url: https://github.com/Serge-sudo/orioledb/sessions/65fe238f-3331-4629-bc9c-a7aad6d13c56

+++ src/transam/predlock.c
@@ -333,7 +333,7 @@ lexicographic_key_gap(OPredLockKey *left, OPredLockKey *right)
 	{
 		size_t		len_diff = right->len - left->len;
 
-		return len_diff > INT_MAX ? INT_MAX : (int) len_diff;
+		return len_diff > (size_t) INT_MAX ? INT_MAX : (int) len_diff;
 	}
 
 	/* Equal-length identical keys: treat as zero additional gap. */
